Move button multiplexer to satisfy deps
