+++ libs/core/pxt.h
@@ -113,4 +113,7 @@ typedef jacdac::JDProxyDriver* JacDacDriverStatus;
 #define DEVICE_ID_FIRST_BUTTON 4000
 #define DEVICE_ID_FIRST_TOUCHBUTTON 4100
 
+#define PXT_INTERNAL_KEY_UP 2050
+#define PXT_INTERNAL_KEY_DOWN 2051
+
 #endif
